+++ Sora/PackageInfo.swift
@@ -9,16 +9,16 @@ public enum SDKInfo {
  */
 public enum WebRTCInfo {
     /// WebRTC フレームワークのバージョン
-    public static let version = "M116"
+    public static let version = "M119"
 
     /// WebRTC フレームワークのコミットポジション
-    public static let commitPosition = "6"
+    public static let commitPosition = "2"
 
     /// WebRTC フレームワークのメンテナンスバージョン
     public static let maintenanceVersion = "1"
 
     /// WebRTC フレームワークのソースコードのリビジョン
-    public static let revision = "44bc8e96ed88005fec89a1cc479e291fea30d1b3"
+    public static let revision = "52bc9f7c1205f4b731ea0289b059f7d240c1e228"
 
     /// WebRTC フレームワークのソースコードのリビジョン (短縮版)
     public static var shortRevision: String {