+++ Sora/Signaling.swift
@@ -2,10 +2,10 @@ import Foundation
 import WebRTC
 
 /// :nodoc:
-private func serializeMetadataList(_ data: Any?) -> [SignalingNotifyMetadata]? {
+private func serializeData(_ data: Any?) -> [SignalingNotifyMetadata]? {
     guard let array = data as? [[String: Any]] else {
         Logger.info(type: .signaling,
-                    message: "downcast failed in serializeMetadataList. data: \(String(describing: data))")
+                    message: "downcast failed in serializeData. data: \(String(describing: data))")
         return nil
     }
 
@@ -77,11 +77,8 @@ private func updateMetadata(signaling: Signaling, data: Data) -> Signaling {
         if let metadata = json["metadata"] {
             message.metadata = metadata
         }
-        if let metadataList = json["metadata_list"] {
-            message.metadataList = serializeMetadataList(metadataList)
-        }
         if let data = json["data"] {
-            message.data = serializeMetadataList(data)
+            message.data = serializeData(data)
         }
         return .notify(message)
     default:
@@ -511,6 +508,9 @@ public struct SignalingNotify {
     /// ロール
     public var role: SignalingRole?
 
+    /// セッション ID
+    public var sessionId: String?
+
     /// クライアント ID
     public var clientId: String?
 
@@ -534,9 +534,6 @@ public struct SignalingNotify {
 
     /// Sora の認証ウェブフックの戻り値で指定された値
     public var authzMetadata: Any?
-
-    /// メタデータのリスト
-    public var metadataList: [SignalingNotifyMetadata]?
 
     /// メタデータのリスト
     public var data: [SignalingNotifyMetadata]?
@@ -569,6 +566,24 @@ public struct SignalingNotify {
 
     /// TURN が利用しているトランスポート層のプロトコル
     public var turnTransportType: String?
+
+    /// 転送フィルターで block または allow となった対象 (audio または video)
+    public var kind: String?
+
+    /// 転送フィルターで block または allow となった送信先の接続 ID
+    public var destinationConnectionId: String?
+
+    /// 転送フィルターで block または allow となった送信元の接続 ID
+    public var sourceConnectionId: String?
+
+    /// 停止された RTP ストリームの送信先接続 ID
+    public var recvConnectionId: String?
+
+    /// 停止された RTP ストリームの送信元接続 ID
+    public var sendConnectionId: String?
+
+    /// 再開された RTP ストリームの送信元接続 ID
+    public var streamId: String?
 }
 
 /**
@@ -1052,6 +1067,7 @@ extension SignalingNotify: Codable {
     enum CodingKeys: String, CodingKey {
         case event_type
         case role
+        case session_id
         case client_id
         case bundle_id
         case connection_id
@@ -1068,6 +1084,12 @@ extension SignalingNotify: Codable {
         case fixed
         case unstable_level
         case turn_transport_type
+        case kind
+        case destination_connection_id
+        case source_connection_id
+        case recv_connection_id
+        case send_connection_id
+        case stream_id
     }
 
     public init(from decoder: Decoder) throws {
@@ -1075,6 +1097,7 @@ extension SignalingNotify: Codable {
         eventType = try container.decode(String.self,
                                          forKey: .event_type)
         role = try container.decodeIfPresent(SignalingRole.self, forKey: .role)
+        sessionId = try container.decodeIfPresent(String.self, forKey: .session_id)
         clientId = try container.decodeIfPresent(String.self, forKey: .client_id)
         bundleId = try container.decodeIfPresent(String.self, forKey: .bundle_id)
         connectionId = try container.decodeIfPresent(String.self,
@@ -1098,6 +1121,18 @@ extension SignalingNotify: Codable {
             try container.decodeIfPresent(Int.self, forKey: .unstable_level)
         turnTransportType =
             try container.decodeIfPresent(String.self, forKey: .turn_transport_type)
+        kind =
+            try container.decodeIfPresent(String.self, forKey: .kind)
+        destinationConnectionId =
+            try container.decodeIfPresent(String.self, forKey: .destination_connection_id)
+        sourceConnectionId =
+            try container.decodeIfPresent(String.self, forKey: .source_connection_id)
+        recvConnectionId =
+            try container.decodeIfPresent(String.self, forKey: .recv_connection_id)
+        sendConnectionId =
+            try container.decodeIfPresent(String.self, forKey: .send_connection_id)
+        streamId =
+            try container.decodeIfPresent(String.self, forKey: .stream_id)
     }
 
     public func encode(to encoder: Encoder) throws {