+++ CHANGES.md
@@ -15,6 +15,9 @@
   - @enm10k
 - [UPDATE] WebRTC 119.6045.2.1 に上げる
   - @miosakuma
+- [UPDATE] VideoView のバックエンドを RTCEAGLVideoView から RTCMTLVideoView に変更する
+  - WebRTC のアップデートに伴い RTCEAGLVideoView が deprecated になったことに伴う修正
+  - @miosakuma
 - [ADD] VideoCodec に H265 を追加する
   - @enm10k
 - [UPDATE] 解像度に qHD (960x540) を追加する