+++ CHANGES.md
@@ -21,6 +21,7 @@
 - [ADD] VideoCodec に H265 を追加する
   - @enm10k
 - [UPDATE] 解像度に qHD (960x540) を追加する
+- [UPDATE] CocoaPods を v1.14.2 に更新する
   - @enm10k
 
 ## 2023.3.1