+++ CHANGES.md
@@ -11,6 +11,10 @@
 
 ## develop
 
+- [CHANGE] SignalingNotify の metadataList を削除する
+  - 2022.1.0 の Sora で metadata_list が廃止されたため
+  - SignalingNotify の data で値の取得が可能
+  - @miosakuma
 - [UPDATE] CameraVideoCapturer のログを出力する
   - @enm10k
 - [UPDATE] WebRTC 119.6045.2.1 に上げる
@@ -27,6 +31,15 @@
 - [UPDATE] ForwardingFilter に version と metadata　を追加する
   - @enm10k @miosakuma
 - [UPDATE] ForwardingFilter の action を未指定にできるようにする
+  - @miosakuma
+- [FIX] SignalingNotify に項目を追加する
+  - sessionId
+  - kind
+  - destinationConnectionId
+  - sourceConnectionId
+  - recvConnectionId
+  - sendConnectionId
+  - streamId
   - @miosakuma
 
 ## 2023.3.1